move list and get-files into their own files
